Moving to version 1.0.0
