+++ tests/spec/NukaCode/Menu/LinkSpec.php
@@ -5,11 +5,6 @@ use Prophecy\Argument;
 
 class LinkSpec extends ObjectBehavior {
 
-    function let()
-    {
-        $this->beConstructedWith('linkName');
-    }
-
     function it_is_initializable()
     {
         $this->shouldHaveType('NukaCode\Menu\Link');
@@ -17,14 +12,8 @@ class LinkSpec extends ObjectBehavior {
 
     function it_checks_set_url()
     {
-        $this->setUrl('testUrl');
+        $this->url = 'testUrl';
         $this->url->shouldBe('testUrl');
-    }
-
-    function it_checks_set_icon()
-    {
-        $this->setIcon('testIcon');
-        $this->icon->shouldBe('testIcon');
     }
 
     function it_checks_set_active()
@@ -39,34 +28,16 @@ class LinkSpec extends ObjectBehavior {
         $this->shouldNotBeActive();
     }
 
-    function it_checks_set_filter_is_false()
-    {
-        $this->setFilter(function(){return true;});
-        $this->shouldNotBeRestricted();
-    }
-
-    function it_checks_set_filter_is_true()
-    {
-        $this->setFilter(function(){return false;});
-        $this->shouldBeRestricted();
-    }
-
-    function it_checks_set_route()
-    {
-        $this->setRoute('test.route');
-        $this->route->shouldBe('test.route');
-    }
-
     function it_checks_set_options()
     {
-        $this->setOptions(['data' => 'dataOne']);
+        $this->options = ['data' => 'dataOne'];
 
         $this->options->shouldContain('dataOne');
         $this->options->shouldHaveKey('data');
     }
 
-    function it_checks_count()
+    function it_checks_is_dropdown()
     {
-        $this->count()->shouldBe(0);
+        $this->isDropDown()->shouldReturn(false);
     }
 }
